backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-jsonfield==1.4.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
jsonfield==3.1.0
PyJWT==2.8.0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
